Service: Manage report on who stores what chunks.
The report contains a list of all known nodes and the hashes
of the chunks they store. This report will be used to make
informed decisions on what chunks to store and speed up chunk
lookups.

This is not need to mine or verify new blocks, so does not
need to be completely accurate.